producer/pyproject.toml
[project]
name = "producer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faststream[cli,kafka]>=0.6.5",
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
